% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{add_iptw}
\alias{add_iptw}
\title{Inverse Probability Treatment Weighting}
\usage{
add_iptw(
  data,
  treatment,
  trt_indicator = 1,
  outcome,
  object = NULL,
  formula = NULL,
  method = c("logit", "rf", "cart", "SVM"),
  mc = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame to be used}

\item{treatment}{Treatment variable name}

\item{trt_indicator}{Value that indicates the unit is treated}

\item{outcome}{Outcome variable name}

\item{object}{A \code{propmod} object if already fitted.}

\item{formula}{If not, write a \link[stats]{formula} to be fitted. Remember that you don't have to worry about group variable. \link[data.table]{.SD} do exclude \code{by}.}

\item{method}{Estimating methods
\itemize{
\item "logit" - \code{\link{ps_glm}}
\item "rf" - \code{\link{ps_rf}}
\item "cart" - \code{\link{ps_cart}}
\item "SVM" - \code{\link{ps_svm}}
}}

\item{mc}{Indicator column name for MC simulation if exists.}

\item{...}{Additional arguments of fitting functions}
}
\description{
fits weighted regression of outcome on treatment
}
\details{
This functions add a column by
\deqn{\frac{A}{\hat{e}_i} + \frac{1- A}{1 - \hat{e}_i}}
where \eqn{A} is the inverse estimated probabilities of treatment.
}
\references{
Pirracchio, R., Petersen, M. L., & Laan, M. van der. (2015). \emph{Improving Propensity Score Estimators’ Robustness to Model Misspecification Using Super Learner}. American Journal of Epidemiology, 181(2), 108–119. \url{https://doi.org/10.1093/aje/kwu253}
}
