% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{add_ipw_wt}
\alias{add_ipw_wt}
\title{Add weighting column for IPW}
\usage{
add_ipw_wt(
  data,
  treatment,
  trt_indicator = 1,
  object = NULL,
  formula = NULL,
  method = c("logit", "rf", "cart", "SVM"),
  mc = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame to be used}

\item{treatment}{Treatment variable name}

\item{trt_indicator}{Value that indicates the unit is treated}

\item{object}{A \code{propmod} object if already fitted.}

\item{formula}{If not, write a \link[stats]{formula} to be fitted. Remember that you don't have to worry about group variable. \link[data.table]{.SD} do exclude \code{by}.}

\item{method}{Estimating methods
\itemize{
\item "logit" - \code{\link{ps_glm}}
\item "rf" - \code{\link{ps_rf}}
\item "cart" - \code{\link{ps_cart}}
\item "SVM" - \code{\link{ps_svm}}
}}

\item{mc}{Indicator column name for MC simulation if exists.}

\item{...}{Additional arguments of fitting functions}
}
\description{
add column before computing IPW and SIPW
}
\details{
This function adds a column named \code{ipw_wt} by
\deqn{\frac{trt_i}{\hat{e}_i} + \frac{1- trt_i}{1 - \hat{e}_i}}
For the usage of the other function, we do not provide any argument for the change of the column name.
}
\seealso{
\code{\link{add_propensity}}
\code{\link{compute_ipw}}
}
