% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{add_propensity}
\alias{add_propensity}
\title{Add estimated propensity score to a data frame}
\usage{
add_propensity(
  data,
  object = NULL,
  formula = NULL,
  method = c("logit", "rf", "cart"),
  var = "propensity",
  ...
)
}
\arguments{
\item{data}{A data frame to be used.}

\item{object}{A \code{propmod} object if already fitted.}

\item{formula}{If not, write a \link[stats]{formula} to be fitted.}

\item{var}{The name of the propensity score column.}

\item{...}{Additional arguments of fitting functions}
}
\description{
adds propensity score to a data frame
}
