% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{build_covariate}
\alias{build_covariate}
\title{Build covariance matrix of the covariates}
\usage{
build_covariate(
  sig = rep(1L, 10),
  confounder_exposure = c(0.2, 0, 0, 0, 0.9, 0, 0, 0, 0, 0, 0, 0),
  confounder_outcome = c(0, 0, 0, 0, 0, 0, 0.2, 0, 0, 0, 0.9, 0),
  exposure_outcome = rep(0L, 9),
  value_cor = TRUE
)
}
\arguments{
\item{sig}{variance of each covariate}

\item{confounder_exposure}{covariance between confounder and exposure predictors, vector in order of column}

\item{confounder_outcome}{covariance between confounder and outcome predictors, vector in order of column}

\item{exposure_outcome}{covariance between exposure predictors and outcome predictors, vector in order of column}

\item{value_cor}{are given values (except \code{sig}) are correlation? (default = TRUE)}
}
\description{
builds covariance matrix of the onfounders, exposure predictors, and outcome predictors
}
\details{
This function builds covariance matrix of the covariates, which are
confounders (w1, w2, w3, w4), exposure predictors (w5, w6, w7), and outcome predictors(w8, w9, w10).
In each group, there is no correlation.
On the other hand, you can specify picewise correlation between e.g. confounder-exposure predictor.
}
\references{
Setoguchi, S., Schneeweiss, S., Brookhart, M. A., Glynn, R. J., & Cook, E. F. (2008). \emph{Evaluating uses of data mining techniques in propensity score estimation: a simulation study}. Pharmacoepidemiology and Drug Safety, 17(6), 546–555 \url{https://doi.org/10.1002/pds.1555}
}
\seealso{
\code{\link{sim_outcome}}
}
