% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{chemical}
\alias{chemical}
\title{Cohort Study Data about POISOX}
\format{
A data frame with 5000 rows and 5 variables:
\describe{
\item{age}{age of each subject}
\item{sex}{sex of each subject, 0 if male, 1 if female}
\item{poisox}{0 if unexposed, 1 if exposed}
\item{mortal}{0 if alive, 1 if dead}
\item{blood}{subsequent blood pressure, after - prior}
}
}
\usage{
chemical
}
\description{
Cohort study interesting in the effect of exposure to an industrial chemical, POISOX, on subsequent blood pressure and mortality.
}
\keyword{datasets}
