% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{compute_asam}
\alias{compute_asam}
\title{Average Standardized Absolute Mean Distance}
\usage{
compute_asam(data, treatment, trt_indicator = 1, outcome)
}
\arguments{
\item{data}{data}

\item{treatment}{column name of treatment}

\item{trt_indicator}{value that indicates the unit is treated, e.g. 1 or TRUE}

\item{outcome}{outcome variable included in the data. It should be specified because it is not covariate.}
}
\description{
computes average standardized absolute mean distance (ASAM)
}
\details{
For each covariate,
compute absolute (standardized difference of means between treatment and control groups), and take average.
Denote that standardization is done by sd of treatment group covariates.

Lower ASAM means that treatment and control groups are more similar w.r.t. the given covariates.
}
\references{
Lee, B. K., Lessler, J., & Stuart, E. A. (2010). \emph{Improving propensity score weighting using machine learning. Statistics in Medicine}. Statistics in Medicine, 29(3), 337-346. \url{https://doi.org/10.1002/sim.3782}
}
