% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{compute_balance}
\alias{compute_balance}
\title{Covariate Balance}
\usage{
compute_balance(
  data,
  col_name = "balance",
  treatment,
  trt_indicator = 1,
  outcome,
  exclude = NULL
)
}
\arguments{
\item{data}{data}

\item{col_name}{column name of standardized difference means between treatment and control groups. By default, "\code{balance}"}

\item{treatment}{column name of treatment}

\item{trt_indicator}{value that indicates the unit is treated, e.g. 1 or TRUE}

\item{outcome}{outcome variable included in the data. It should be specified because it is not covariate.}

\item{exclude}{Additional columns to exlude}
}
\description{
gives covariate balance summary
}
