% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{compute_ipw}
\alias{compute_ipw}
\title{Estimation of Inverse Probability Weighting}
\usage{
compute_ipw(
  data,
  treatment,
  trt_indicator = 1,
  outcome,
  weight = NULL,
  object = NULL,
  formula = NULL,
  method = c("logit", "rf", "cart", "SVM"),
  mc_col = NULL,
  sc_col = NULL,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame to be used}

\item{treatment}{Treatment variable name}

\item{trt_indicator}{Value that indicates the unit is treated}

\item{outcome}{Outcome variable name}

\item{weight}{If weighting column exists, it can be specified for computing efficiency}

\item{object}{A \code{propmod} object if already fitted.}

\item{formula}{If not, write a \link[stats]{formula} to be fitted. Remember that you don't have to worry about group variable. \link[data.table]{.SD} do exclude \code{by}.}

\item{method}{Estimating methods
\itemize{
\item "logit" - \code{\link{ps_glm}}
\item "rf" - \code{\link{ps_rf}}
\item "cart" - \code{\link{ps_cart}}
\item "SVM" - \code{\link{ps_svm}}
}}

\item{mc_col}{Indicator column name for MC simulation if exists}

\item{sc_col}{Indicator column name for various scenarios if exists}

\item{parallel}{parallelize some operation}

\item{...}{Additional arguments of fitting functions}
}
\description{
estimates inverse probability weighting (IPW) based on propensity score estimates
}
