% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\docType{data}
\name{outcome_frame}
\alias{outcome_frame}
\title{Potential Outcome Framework Toy Dataset}
\format{
A data frame with 8 rows and 5 variables:
\describe{
\item{y0}{outcome when the unit is untreated}
\item{y1}{outcome when the unit is treated}
\item{z}{binary treatment}
\item{y}{outcome}
\item{n}{observed count}
}
}
\usage{
outcome_frame
}
\description{
Toy example for potential outcome framework. The table construction is given by:
}
\keyword{datasets}
