% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{plot_cp}
\alias{plot_cp}
\title{Plot a Complexity Parameter Table}
\usage{
plot_cp(object, ...)
}
\arguments{
\item{object}{fitted \code{\link{propmod}} object. Its \code{name} should be "cart"}

\item{...}{Additional arguments for \link[ggplot2]{geom_pointrange}}
}
\description{
Before pruning, plot complexity parameter from CART.
}
