% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensityml-package.R
\docType{package}
\name{propensityml-package}
\alias{propensityml-package}
\alias{_PACKAGE}
\alias{propensityml}
\title{propensityml: Propensity score weighting using machine learning methods}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Implement machine learning models to improve propensity score weight,
    especially tree models, e.g.
    CART, pruned CART, random forests, etc.
}
\details{
The propensityml package provides various ML functions to estimate propensity score.
}
\references{
Lee, B. K., Lessler, J., & Stuart, E. A. (2010). \emph{Improving propensity score weighting using machine learning. Statistics in Medicine}. Statistics in Medicine, 29(3), 337-346.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ygeunkim}
}

}
\author{
\strong{Maintainer}: Young Geun Kim \email{dudrmshub@gmail.com}

}
\keyword{internal}
