% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propmod-class}
\alias{propmod-class}
\alias{propmod}
\alias{cvar-class}
\alias{print.propmod}
\title{\code{propmod} class}
\usage{
\method{print}{propmod}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\description{
The \code{propmod} class ready for computing propensity score
}
