% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{ps_cart}
\alias{ps_cart}
\title{Fitting CART for Propensity Score}
\usage{
ps_cart(formula, data, ...)
}
\arguments{
\item{formula}{an object \link[stats]{formula} to be fitted. Response should be treatment.}

\item{data}{optional data frame.}

\item{...}{For additional options of \link[rpart]{rpart}.}
}
\value{
\code{\link{propmod}} class, a list with model and its name
\itemize{
\item model - \link[rpart]{rpart}
\item name - "cart"
\item data - dataset
}
}
\description{
fits classification tree before estimating propensity scores
}
\references{
Lee, B. K., Lessler, J., & Stuart, E. A. (2010). \emph{Improving propensity score weighting using machine learning. Statistics in Medicine}. Statistics in Medicine, 29(3), 337-346. \url{https://doi.org/10.1002/sim.3782}
}
