% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{ps_glm}
\alias{ps_glm}
\title{Fitting Logistic Regression for Propensity Score}
\usage{
ps_glm(formula, data, ...)
}
\arguments{
\item{formula}{an object \link[stats]{formula} to be fitted. Response should be treatment.}

\item{data}{optional data frame.}

\item{...}{For additional options of \link[stats]{glm}.}
}
\value{
\link[propensityml]{propmod} class, a list with model and its name
\itemize{
\item model - \link[stats]{glm} model
\item name - "glm"
\item data - dataset
}
}
\description{
fits logistic regression model before estimating propensity scores
}
\details{
Propensity score is
\deqn{e(X) = P(Z_i = 1 \mid X_i = x)},
which is the conditional probability of receiving treatment.
Naturally, logit model is the easiest way to estimate the score.
}
\references{
Rosenbaum, P. R., & Rubin, D. B. (1983). \emph{The central role of the propensity score in observational studies for causal effects}. Biometrika, 70(1), 41-55. \url{https://doi.org/10.1093/biomet/70.1.41}
}
