% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{ps_prune}
\alias{ps_prune}
\title{Pruning CART}
\usage{
ps_prune(object, cp, ...)
}
\arguments{
\item{object}{fitted \code{\link{propmod}} object. Its \code{name} should be "cart"}

\item{...}{Additional arguments for \link[rpart]{prune}.}
}
\description{
prunes the result of \code{\link{ps_cart}}.
}
