% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{ps_svm}
\alias{ps_svm}
\title{Fitting SVM for Propensity Score}
\usage{
ps_svm(
  formula,
  data,
  scale = FALSE,
  kernel = c("radial", "linear", "polynomial", "sigmoid"),
  cost = 1,
  gamma,
  ...
)
}
\arguments{
\item{formula}{an object \link[stats]{formula} to be fitted. Response should be treatment.}

\item{data}{optional data frame.}

\item{scale}{scale the data?}

\item{kernel}{kernel used in SVM. See \link[e1071]{svm}}

\item{cost}{cost parameter C}

\item{gamma}{parameter in some kernel}

\item{...}{For additional options of \link[e1071]{svm}}
}
\value{
\code{\link{propmod}} class, a list with model and its name
\itemize{
\item model - \link[e1071]{svm}
\item name - "SVM"
\item data - dataset
}
}
\description{
Fits Support Vector Machine before estimating propensity scores
}
