% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{sim_covariate}
\alias{sim_covariate}
\title{Simulating covariates}
\usage{
sim_covariate(n, covmat = build_covariate())
}
\arguments{
\item{n}{sample size}

\item{covmat}{Covariance matrix of the covariates}
}
\description{
generates covariates in the references
}
\details{
This function reproduces the setting in the paper in Setoguchi et al. and Lee et al.
First generate binary covariates (w1, w3, w5, w6, w8, w9), and continuous covariates (w2, w4, w7, w10).
\enumerate{
\item Generate 10-dim multivariate normal v1, v3, v5, v6, v8, v9 (corresponding to binary), and w7, w10
}
}
\references{
Setoguchi, S., Schneeweiss, S., Brookhart, M. A., Glynn, R. J., & Cook, E. F. (2008). \emph{Evaluating uses of data mining techniques in propensity score estimation: a simulation study}. Pharmacoepidemiology and Drug Safety, 17(6), 546–555 \url{https://doi.org/10.1002/pds.1555}

Lee, B. K., Lessler, J., & Stuart, E. A. (2010). \emph{Improving propensity score weighting using machine learning. Statistics in Medicine}. Statistics in Medicine, 29(3), 337-346. \url{https://doi.org/10.1002/sim.3782}
}
\seealso{
\code{\link{build_covariate}}
}
