% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{add_iptw}
\alias{add_iptw}
\title{Treatment Effect Estimation Using Propensity Scores}
\usage{
add_iptw(
  data,
  treatment,
  trt_indicator = 1,
  object = NULL,
  formula = NULL,
  method = c("logit", "rf", "cart", "SVM"),
  mc_col = NULL,
  sc_col = NULL,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame to be used}

\item{treatment}{Treatment variable name}

\item{trt_indicator}{Value that indicates the unit is treated}

\item{object}{A \code{propmod} object if already fitted.}

\item{formula}{If not, write a \link[stats]{formula} to be fitted. Remember that you don't have to worry about group variable. \link[data.table]{.SD} do exclude \code{by}.}

\item{method}{Estimating methods
\itemize{
\item "logit" - \code{\link{ps_glm}}
\item "rf" - \code{\link{ps_rf}}
\item "cart" - \code{\link{ps_cart}}
\item "SVM" - \code{\link{ps_svm}}
}}

\item{mc_col}{Indicator column name for MC simulation if exists}

\item{sc_col}{Indicator column name for various scenarios if exists}

\item{parallel}{parallelize some operation}

\item{...}{Additional arguments of fitting functions}
}
\description{
estimates treatment effect based on ps estimation (e.g. inverse probability treatment weighting)
}
\details{
This functions add columns by
\deqn{\frac{trt_i}{\hat{e}_i} + \frac{1- trt_i}{1 - \hat{e}_i}}
and
\deqn{trt_i + (1 - trt_i) \frac{\hat{e}_i}{1 - \hat{e}_i}}
}
\references{
Lee, B. K., Lessler, J., & Stuart, E. A. (2010). \emph{Improving propensity score weighting using machine learning. Statistics in Medicine}. Statistics in Medicine, 29(3), 337-346.

Pirracchio, R., Petersen, M. L., & Laan, M. van der. (2015). \emph{Improving Propensity Score Estimators’ Robustness to Model Misspecification Using Super Learner}. American Journal of Epidemiology, 181(2), 108–119. \url{https://doi.org/10.1093/aje/kwu253}
}
\seealso{
\code{\link{add_propensity}}
}
