% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{compute_asam}
\alias{compute_asam}
\title{Average Standardized Absolute Mean Distance}
\usage{
compute_asam(
  data,
  treatment,
  trt_indicator = 1,
  outcome,
  exclude = NULL,
  object = NULL,
  formula = NULL,
  method = c("logit", "rf", "cart", "SVM"),
  weighting = c("IPW", "SIPW"),
  mc = NULL,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{data}{data}

\item{treatment}{column name of treatment}

\item{trt_indicator}{value that indicates the unit is treated, e.g. 1 or TRUE}

\item{outcome}{outcome variable included in the data. It should be specified because it is not covariate.}

\item{exclude}{Additional columns to exlude}

\item{object}{A \code{propmod} object if already fitted.}

\item{formula}{If not, write a \link[stats]{formula} to be fitted. Remember that you don't have to worry about group variable. \link[data.table]{.SD} do exclude \code{by}.}

\item{method}{Estimating methods
\itemize{
\item "logit" - \code{\link{ps_glm}}
\item "rf" - \code{\link{ps_rf}}
\item "cart" - \code{\link{ps_cart}}
\item "SVM" - \code{\link{ps_svm}}
}}

\item{weighting}{Weighting methods, IPW or SIPW}

\item{mc}{Indicator column name for group if exists, e.g. c("mcname", "scenario")
\itemize{
\item MC simulation
\item Scenario
}}

\item{parallel}{parallelize some operation}

\item{...}{Additional arguments of fitting functions}
}
\description{
computes average standardized absolute mean distance (ASAM)
}
\details{
For each covariate,
compute absolute (standardized difference of means between treatment and control groups), and take average.
Denote that standardization is done by sd of treatment group covariates.

Lower ASAM means that treatment and control groups are more similar w.r.t. the given covariates.
}
\references{
Lee, B. K., Lessler, J., & Stuart, E. A. (2010). \emph{Improving propensity score weighting using machine learning. Statistics in Medicine}. Statistics in Medicine, 29(3), 337-346. \url{https://doi.org/10.1002/sim.3782}
}
