% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{compute_ipw}
\alias{compute_ipw}
\title{Estimation of Inverse Probability Weighting}
\usage{
compute_ipw(
  data,
  treatment,
  trt_indicator = 1,
  outcome,
  object = NULL,
  formula = NULL,
  method = c("logit", "rf", "cart", "SVM"),
  ...
)
}
\arguments{
\item{data}{A data frame to be used}

\item{treatment}{Treatment variable name}

\item{trt_indicator}{Value that indicates the unit is treated}

\item{outcome}{Outcome variable name}

\item{object}{A \code{propmod} object if already fitted.}

\item{formula}{If not, write a \link[stats]{formula} to be fitted.}

\item{method}{Estimating methods
\itemize{
\item "logit" - \code{\link{ps_glm}}
\item "rf" - \code{\link{ps_rf}}
\item "cart" - \code{\link{ps_cart}}
\item "SVM" - \code{\link{ps_svm}}
}}

\item{...}{Additional arguments of fitting functions}
}
\description{
estimates inverse probability weighting (IPW) based on propensity score estimates
}
