% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{estimate_ps}
\alias{estimate_ps}
\title{Estimation of Propensity Score}
\usage{
estimate_ps(object, ...)
}
\arguments{
\item{object}{fitted \link[propensityml]{propmod} object}

\item{...}{additional arguments for \link[stats]{predict}}
}
\description{
estimates propensity score based on the given model
}
